[reconfig] Add certificate creation trait (#6806)
* [AuthAgg] Add auth agg creation trait

* Reorganize tests
