+++ crates/sui-core/src/test_utils.rs
@@ -2,9 +2,25 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::authority::{AuthorityState, EffectsNotifyRead};
+use crate::authority_aggregator::{AuthorityAggregator, TimeoutConfig};
+use crate::authority_client::LocalAuthorityClient;
+use crate::epoch::committee_store::CommitteeStore;
+use fastcrypto::traits::KeyPair;
+use prometheus::Registry;
 use signature::Signer;
+use std::collections::BTreeMap;
+use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
+use sui_config::genesis::Genesis;
+use sui_config::ValidatorInfo;
+use sui_framework_build::compiled_package::BuildConfig;
+use sui_types::base_types::ObjectRef;
+use sui_types::crypto::AuthorityKeyPair;
+use sui_types::crypto::{
+    generate_proof_of_possession, get_key_pair, AccountKeyPair, AuthorityPublicKeyBytes,
+    NetworkKeyPair, SuiKeyPair,
+};
 use sui_types::utils::create_fake_transaction;
 use sui_types::{
     base_types::{random_object_ref, AuthorityName, ExecutionDigests, TransactionDigest},
@@ -13,7 +29,7 @@ use sui_types::{
     gas::GasCostSummary,
     message_envelope::Message,
     messages::{CertifiedTransaction, ExecutionStatus, Transaction, TransactionEffects},
-    object::Owner,
+    object::{Object, Owner},
 };
 use tokio::time::timeout;
 use tracing::{info, warn};
@@ -100,4 +116,114 @@ pub fn dummy_transaction_effects(tx: &Transaction) -> TransactionEffects {
         events: Vec::new(),
         dependencies: Vec::new(),
     }
+}
+
+async fn init_genesis(
+    committee_size: usize,
+    mut genesis_objects: Vec<Object>,
+) -> (
+    Genesis,
+    Vec<(AuthorityPublicKeyBytes, AuthorityKeyPair)>,
+    ObjectRef,
+) {
+    // add object_basics package object to genesis
+    let build_config = BuildConfig::default();
+    let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+    path.push("src/unit_tests/data/object_basics");
+    let modules = sui_framework::build_move_package(&path, build_config)
+        .unwrap()
+        .get_modules()
+        .into_iter()
+        .cloned()
+        .collect();
+    let pkg = Object::new_package(modules, TransactionDigest::genesis()).unwrap();
+    let pkg_ref = pkg.compute_object_reference();
+    genesis_objects.push(pkg);
+
+    let mut builder = sui_config::genesis::Builder::new().add_objects(genesis_objects);
+    let mut key_pairs = Vec::new();
+    for i in 0..committee_size {
+        let key_pair: AuthorityKeyPair = get_key_pair().1;
+        let authority_name = key_pair.public().into();
+        let worker_key_pair: NetworkKeyPair = get_key_pair().1;
+        let worker_name = worker_key_pair.public().clone();
+        let account_key_pair: SuiKeyPair = get_key_pair::<AccountKeyPair>().1.into();
+        let network_key_pair: NetworkKeyPair = get_key_pair().1;
+        let validator_info = ValidatorInfo {
+            name: format!("validator-{i}"),
+            protocol_key: authority_name,
+            worker_key: worker_name,
+            account_key: account_key_pair.public(),
+            network_key: network_key_pair.public().clone(),
+            stake: 1,
+            delegation: 0,
+            gas_price: 1,
+            commission_rate: 0,
+            network_address: sui_config::utils::new_tcp_network_address(),
+            p2p_address: sui_config::utils::new_udp_network_address(),
+            narwhal_primary_address: sui_config::utils::new_udp_network_address(),
+            narwhal_worker_address: sui_config::utils::new_udp_network_address(),
+        };
+        let pop = generate_proof_of_possession(&key_pair, (&account_key_pair.public()).into());
+        builder = builder.add_validator(validator_info, pop);
+        key_pairs.push((authority_name, key_pair));
+    }
+    let genesis = builder.build();
+    (genesis, key_pairs, pkg_ref)
+}
+
+pub async fn init_local_authorities(
+    committee_size: usize,
+    genesis_objects: Vec<Object>,
+) -> (
+    AuthorityAggregator<LocalAuthorityClient>,
+    Vec<Arc<AuthorityState>>,
+    ObjectRef,
+) {
+    let (genesis, key_pairs, pkg_ref) = init_genesis(committee_size, genesis_objects).await;
+    let (aggregator, authorities) = init_local_authorities_with_genesis(&genesis, key_pairs).await;
+    (aggregator, authorities, pkg_ref)
+}
+
+pub async fn init_local_authorities_with_genesis(
+    genesis: &Genesis,
+    key_pairs: Vec<(AuthorityPublicKeyBytes, AuthorityKeyPair)>,
+) -> (
+    AuthorityAggregator<LocalAuthorityClient>,
+    Vec<Arc<AuthorityState>>,
+) {
+    telemetry_subscribers::init_for_testing();
+    let committee = genesis.committee().unwrap();
+
+    let mut clients = BTreeMap::new();
+    let mut states = Vec::new();
+    for (authority_name, secret) in key_pairs {
+        let client = LocalAuthorityClient::new_with_objects(
+            committee.clone(),
+            secret,
+            genesis.objects().to_owned(),
+            genesis,
+        )
+        .await;
+        states.push(client.state.clone());
+        clients.insert(authority_name, client);
+    }
+    let timeouts = TimeoutConfig {
+        authority_request_timeout: Duration::from_secs(5),
+        pre_quorum_timeout: Duration::from_secs(5),
+        post_quorum_timeout: Duration::from_secs(5),
+        serial_authority_request_timeout: Duration::from_secs(1),
+        serial_authority_request_interval: Duration::from_secs(1),
+    };
+    let committee_store = Arc::new(CommitteeStore::new_for_testing(&committee));
+    (
+        AuthorityAggregator::new_with_timeouts(
+            committee,
+            committee_store,
+            clients,
+            &Registry::new(),
+            timeouts,
+        ),
+        states,
+    )
 }
